test(profilingOnboardingModal): fix test
